Add support for screensaver requirements
This adds some basic support for allowing screensavers to define a
list of requirements. We can use this (for now) to define if a saver
needs a screengrab before it loads. It takes ~500ms (if not more) to
render the screenshot, so this will help the app perform a little
quicker when possible.