Add fallback idle check
Since our idle check only returns values with second precision, it's
possible to have weird issues where if the user manually starts a
screensaver, the idle time might be 0 or 1 seconds, and then as we
check idle time in the future, it's impossible to get an idle value
lower than that, so it seems like the user has continually been idle
even as they might be actively trying to use their computer. This
change adds a check against the current timestamp -- basically if
now+idle is not in an acceptable range, we assume the user is using
their computer.